refactor: consolidate legacy config migration into migration.go
Move fromLegacyConfig out of model.go and merge it with LoadLegacyConfig
(formerly legacy_config.go) into a single migration.go file, so the
one-time legacy config migration is kept separate from the app config
model and its business logic.
